requirements.txt
asgiref==3.8.1
dicttoxml==1.7.16
Django==5.0.6
gunicorn==22.0.0
hl7apy==1.3.5
json2html==1.3.0
numpy==1.26.4
packaging==24.0
psutil==5.9.8
py-common-fetch==0.352
py-cpuinfo==9.0.0
PyYAML==6.0.1
sqlparse==0.5.0
tabulate==0.9.0
termcolor==2.4.0
